Largely rewritten.  Should now handle SQLite dynamic field types.
Still a lot of work needing to be done, but this should at least
display arbitrary SQLite databases without complete barfage.

Something to be aware of is the effective disabling of $sce
escaping on the displayed field output, which will very probably
need changing.  It likely allows direct display of HTML embedded
in user databases, which would allow for XSS (and likely worse).

One potential alternative might be to use a custom AngularJS
filter instead, which would properly escape all output except for
'&nbsp;' which we want to keep (to stop whitespace representing
empty fields from being collapsed).