Use user name in subject CN in client cert generation
The subject CN (as opposed to the issuer CN) is more of a certificate
owner. In the case of client certs this is probably the user name (or
the user name plus something). So this should be more correct. It also
allows the user to distinguish between multiple client certs for the
same host.