Ensure database page shows correct star and fork count
Fixes #9.  Previously, we were caching all of the database stats in
memcached.

This was a problem when the star and/or fork count changed after
the page was cached, as the old star/fork counts would be shown.

With this commit, the latest star and fork counts are retrieved
from the database separately, after the bulk of the page is
retrieved from memcached.