Make discussion id generation more robust
Hit the limits of the previous super simplistic id generation
approach.  This one should be better.  It's still open to a race
condition, but the window for that seems like it would take a lot
of heavy contention on the database for that to occur.  Not sure
how likely that is in the short/medium term future, so this is
good enough for now.