Potential fix for the branch name duplication problem on db page
The underlying cause isn't readily apparent though, as the problem
data is coming from memcache.  The only code path which sets the
data seems like this shouldn't happen, so the fix here is just a
guess.  Also added code to automatically check for the problem when
pulling data from memcache, then if found log a warning so an admin
can investigate.