Add checks and handling for the default table not existing
This could (previously) happen when a new commit has different
table in it, with the previously chosen default now missing.

Similarly, this could also happen when deleting a commit, as the
new head commit might have the default table missing.

Whenever this happened, an error message was displayed instead of
something useful, thereby rendering the database almost inaccessible.

This commit adds catches and handling for this, clearing the
old/invalid default table.

This commit also adds dynamic table lookup to the settings page,
so when choosing a different branch the tables for the new branch
are shown.