Use <a> tags with href attributes instead of <button> for some links
Button tags don't include a href which can be copied and used by
browsers, making them a bit of a pain to use.  This is especially
important with the release download links.  This commit changes the
release and "New Merge Request" buttons so they not have href links
which can be copied and pasted using right click browser context
menu.  eg "Copy Link Location"