Save point.  Testing concepts for conditionally displaying the vis canvas
This commit uses Go template directives to hide elements.  Works for
not showing the data, but as the elements aren't in the DOM at all,
they can't be "enabled" real-time after data has been fetched.

Probably better to try something on the Angular side, for dynamic
suppression/enabling of the canvas depending on the data being available
to render.