webui vis: don't reset columns after retrieving data for saved vis
This was a pita, as I still don't seem to understand how to execute
a function with a http.get() inside it, in a blocking fashion.

eg so things execute in the stated order, instead of in some weird
parallel race-condition happy way. :(

As a workaround, had to duplicate most of the existing execSQL()
function that otherwise could have been shared code. :(