db4s: Use RFC3339 for timestamps everywhere
The commit list in the metadata is using RFC3339 for timestamps while
the user and database lists use RFC822. This makes parsing the resulting
strings needlessly complicated. Because changing the timestamps in the
metadata seems more complicated, this commit changes all usages of
RFC822 to RFC3339. This should not cause any problems because the only
program using this data is DB4S which currently only shows the string
as-is.