db4s: Replace "Upgrade required" error message
Replace the old "Upgrade required" error message when trying to push an
existing database without providing a commit id. The message aimed at
very old versions of DB4S which are not much used anymore. It can
however show up in regular use, e.g. when trying to push a new database
to an existing name, and is very confusing in these situations.