webui: Very simple implementation of copying a cloning link for DB4S
This adds a button for copying a link which allows cloning the current
database and branch in DB4S. The current look and feel of this feature
is far from perfect. Ideally we would show a nicer popup dialog with
some extra information like a download link to DB4S and a note to
remember using a client certificate.