api: Add diff handler to v1 API
This adds a /v1/diff handler which calls the com.Diff function and
returns its results in JSON format. While this might be useful later its
main purpose at the moment is to expose the com.Diff function in an easy
to call way for testing purposes. Because of this the API handler is not
added to the help page yet.