common: Include changed table data in diff results
This adds diffing of table data and includes the changed rows in the
diff results. Note that while this tells you which rows have changed it
does not yet include the SQL statements required for merging the data.

This also makes the schema diff optional. If there are no changes to the
schema, the resulting schema diff will be nil.