common: Fix conversion of numbers to strings when reading rows
When reading in the results of a query numbers are converted to strings
for later use. For integers we want to use 64bit integers internally to
make sure every possible integer fits in our variables. For floats we
want to make sure we use the highest required precision when converting
the number to a string instead of rounding early.