api: Also allow authentication using client certificate
This allows authentication using the same client certificates as the
DB4S end point. When an API key is provided this is still preferred.
This change not only allows regular users of the API to use another
authentication mechanism but it also allows DB4S to use the API calls
which eventually might allow us to drop some functions from the db4s
code and move it to the api code.