api: Add branches and metadata handlers
Add two new handlers to the API: branches and metadata. The first only
returns a list of all branches while the second returns a list of all
branches, commits, releases, and tags.

The output from both functions is the same as the respective output from
the functions in the db4s end point.