webui: Add stub for diff page
Add a new page for showing the differences between two commits of the
same database. Currently this page only lists the name of the objects
which changes along with some rudimentary statistics.

Add a new button to the commit list page which redirects you to the new
differences page showing the changes made in this commit.