common: Sort diff changesets by object type
Sorting the changesets in the Diff function by type allows us to execute
the generated SQL statements in the returned order without any problems
because we now make sure to always change tables first and views
triggers. This is because everything depends on the table schemas while
triggers can also depend on a view schema but not vice versa.