common, webui: Reduce size of Diff() output
This reduces the size of the output of the Diff() function a lot by
removing the column names and data types in the row data arrays and
omitting empty attributes in the JSON output.

This requires the webUI code to be changed to deal with the changed
data format. The reduced size of the JSON data should speed up the web
page noticeably for larger diffs.