webui: Add button to compare branches to branch list
This adds a new button called "Compate with [default branch name]" to
each branch in the branches list. Clicking this button redirects to the
diff page comparing the default branch with that branch.

Also make the Delete branch button red and rename the Update button to
"Save Changes" which I find less misleading.