common: Perform conflict check when fast.forward merge is not possible
When a fast-forward merge is not possible and we need to perform an
actual merge conflicts can happen. Since we have no strategy in place to
solve these conflicts we need to check for them and abort the merge if
there is one.

Because the actual merge is not yet implemented even in the case where
there is no conflict this change effectively only adds another possible
error message at the moment: either there is a conflict which makes the
merge impossible or the merge is possible but not yet implemented.