common: Implement the actual merging of databases
This adds the actual merging of branches which cannot be simply fast
forwarded. The resulting database file should contain the changes of
both branches since the last common ancestor, except for those cases
where there are conflicts. The merged database is not stored yet nor do
we create a merge commit. So from the user point of view nothing
changes.

Also simplify the code which checks for conflicts a bit.