Add database sharing
This adds a simple way to share databases between users. It allows you
to give other users read and/or write access to your public and private
databases on the Settings page. Read access is useful for sharing
private database, write access is useful for collaborative work. These
settings are respected by the webUI, the API and the DB4S module.

More finegrained control over permissions might be useful in the future.
Also shared private databases do not show up on the webUI nor in the
DB4S listings.