common: Fix memcached key generation
The keys used for storing data in memcached can cause problems: Because
all users who are not the owner of the database generate the same key,
they see the same cached version of the page. This causes a problem when
a logged in user sees a database they do not own which produces a cache
item which includes the user's avatar and similar. A user who is not
logged in then would get this user's version of the page.