webui: Add another button for generating client certs to settings page
Add a second button for generating the client certificates for using
DB4S or dio to the webui. There was only one one the main profile page
but not on the settings page. Since the settings page is at least as
obvious to reach when working on some database and since it already
manages the API keys we might as well add a second button there.