webui: Fix a91ed9ffd468368ff945da06c89e156c8ed243bd
This fixes some issues introduced in a previous commit for deduplicating
code for live databases. The issue here was that the common database
header now relies on an information whether the current database is a
live database. For mose pages this information was not yet provided.
Instead of providing this information for all pages by adding more code,
this commit moves the deduplication forward by updating the existing
functions to include this information.