webui: Refactor data passed to web page templates
This refactors the page data structures passed to the HTML templates.
The idea here is to have one data structure for meta information on the
web page, including authentication and server information, and one
separate data structure on the database.

By improving the helper functions for handling these structures this
also simplifies the code a bit.