webui: Refactor special case of getting owner & database in upload page
The upload page gets the database owner and name via GET parameters
instead of extracting them from the actual URL. Also, unlike the other
pages which extract them from the URL, not providing this information is
permitted here. Because this is a unique case move the code to the
uploadPage function instead of having it in the generic function.