webui: Move code for getting database owner and name to own function
So far the code for getting the database owner and name was included in
the function for getting page meta information. Move that code to a
separate function that is only called for pages dealing with databases.
This makes it more obvious what is happening.

Note that in the new function, unlike before, permissions are no longer
checked. It only handles data retrieval. This is however ok because in
all cases where it is used permissions are checked later anyway, usually
by calling com.CheckDBPermissions or com.DBDetails.