webui: Move retrieving star and watch status to DBDetails function
Instead of querying the watch and star status of a database in each web
page manually, add the code to the DBDetails function. Basically every
call of the DBDetails function is followed by querying the watch and
star status anyway. Arguably this information is part of the database
details anyway because they are always presented from the perspective of
the currently logged in user.