webui: Refactor branches page
Rewrite the entire branches page. The Go code for retrieving the data is
simplified a lot, the Go HTML template is minimised as much as possible,
and all the rendering and frontend handling is moved to React. This
includes some minor changes to how the UI looks but no major breaks.

The code for the markdown editor component is slightly extended to
accomodate the needs by the new branches code.