webui: Refactor database page
Rewrite the database page to use React instead of AngularJS. While doing
so this cleans up the existing code, especially on the frontend side, a
lot. The looks of the page change slightly but the functionality is
mostly untouched. Only resizing of the table columns is added.

Note that this adds jQuery as a dependency. This is just temporarily.
Please do not start using it.

This also changes how NULL values are passed from the server to the
frontend. This means that the memcache has to be cleared for NULL values
to show up correctly in the frontend.