webui: Add basic updating of tables to database page for live databases
This adds basic update functionality to the database page of live
databases. By double clicking a cell a text editor opens and a new value
can be typed in. Copy and paste of single cell values works as well.

For this feature to work a list of the primary key columns must be
provided by the server. This list is also used in the code for non-live
databases to be able to distinguish rows. So ideally the memcache should
be cleared when rolling out these changes or else the added information
is not provided for cached pages.