api, common, live, webui:  Use a generated bucket name and id for live databases
With this commit (which needs some database fields added) we move to
using randonly generated strings for the Minio bucket name and
object id when storing live databases.

The randomly generated string used as the bucket name is kept in
each users database record, so all of a users live databases will
be stored in the same directory in the Minio backend.  The object
id is uniquely generated for each database.

Manual PostgreSQL database changes needed for this commit:

    alter table public.users add live_minio_bucket_name text;
    alter table public.sqlite_databases add live_minio_object_id text;