webui: Improve database view row buttons and honour user maxRows preference
This adds the database paging controls to the top of the data grid
as well, as having them at both the top and bottom is much more
convenient than just at the bottom when viewing large numbers of
rows.

Additionally, this commit replaces the dodgy existing unicode based
control icons (that were bright orange in Chromium) with font awesome
icons. They needed css borders added for visual separation, but the
result seems ok.

The row count numbers have been moved to the right of the arrows as
well, to help make the button locations not jump around as much when
trying to click through several pages.  This bit still needs
improvement, so the button locations don't move at all.

The data grid also now honours the user's selected "maximum number
of rows to display" setting, manually setting the grid to that size.