webui: Migrate profile page to React
Besides migrating this page to React this also updates the common code
for getting a list of the live databases of a user easier to use. Also
the information shown in the profile and in the user pages is more
similar. The same is true for the information on standard and on live
databases. This also means that download and view count for public
databases are now also visible to other users. Please note though that
this is *not* a change in the permissions as this information has been
sent to the client before too - it just was not shown in the frontend.
Additionally this removes the "live" label from shared live databases
and removed the "beta testing" label from them as well. The idea here is
to a) not put too much effort into the profile page because it should
probably be redesigned anyway and b) start working towards making live
and standard databases more similar. Finally this commit removes the
"Expand/Collapse" buttons and replaces them by individual toggle buttons
for each database.