webui: Complete redesign of the execute SQL page
This migrates the execute SQL page to React. While doing so the design
and features are changed a lot.

The layout vaguely resembles a terminal session now which means that you
see the history of the last executed SQL statements as you type in the
next statement. The idea is to differentiate the execute and the
visualise page by making the execute page aim more at modifying
statements (no need to be able to save them under a name most of the
time, multiple in a row are common, feedback what was already done is
helpful).

Besides showing a command history also the editor is improved. It now
features a better font as well as syntax highlighting. It auto focuses
when the page is loaded so you can just start typing. There is also a
feature to execute the commands using the keyboard (including by just
pressing the Enter key), making this more similar to using a shell. Less
common options like the format SQL button are moved to a menu.

This drops the concept of saved SQL statements for the execute page. The
database migrations are therefore a bit more drastic:
 DROP TABLE live_user_sql_statements;
Besides not being used in the current frontend approach anymore, the
live_user_sql_statements table also had no column to save statements per
database. In many contexts however, applying a statement between
different databases (and also on a regular basis) seems pretty rare.