common, cypress: Change our PG library to jackc/pgx/v5
We've been using v3 so far, which was working well,
however its not supported by golang-migrate.

Initial impression of this new version is that it seems
needlessly complex. Although there's some documentation,
there doesn't seem to be much in the way of useful
examples.  There are no upgrade docs (eg "these new
types replace the old ones..."), nor release notes
(not even for major versions), etc. :/

So, this is definitely a "change", rather than an
"upgrade".