database, common: Switch to using migrations instead of manual db changes
This is super overdue. ;)

With this commit, we switch to using automatic database migrations
instead of needing to manually apply database changes.

To manage the migrations, we're using golang-migrate:

  https://github.com/golang-migrate/migrate

New migrations are automatically applied by our daemons, whenever
any of them connects to our PostgreSQL backend.

New migrations should be created in the database/migrations folder.