Use Erlang and RabbitMQ from the official RabbitMQ docker image
We do it this way because the previous approach (using erlang
and rabbitmq-server Alpine packages) was unreliable.

Sometimes with the previous approach, the version of Erlang
available via apk wouldn't match the version RabbitMQ was compiled
with, causing RabbitMQ to fail.

With this new approach, we should always have matching Erlang and
RabbitMQ, so it should all work ok.