webui: Allow embedding visualisations on external sites
Add a new read-only way of rendering visualisations without any of the
editor and page layout around it, allowing for embedding the
visualisation on external pages using an iframe.

Also add a new button in the visualisation editor which produces the
HTML code the user needs to embed the chart.