webui: Update Bootstrap to version 5.3.2
Upgrade Bootstrap from version 3.4.1 to version 5.3.2 (i.e. the latest
version at the moment). This is a major change, upgrading by two major
releases, to it requires a lot of adjustments in the current templates
and HTML code.

While adjusting the existing markup to be compatible with Bootstrap 5,
also rewrite parts of the markup to make better use of semantically
correct tags (e.g. not use tables for layouting only), reduce the use of
inline CSS styles (and use the proper Bootstrap classes instead), unify
the look and feel of the web UI a bit, and fix some HTML validation
errors.

Dependency-wise this removed the Glyphicon and the jQuery dependencies
completely. Bootstrap itself is moved to the webpack bundle instead of
being a set of separate files.