api: Use Gin for routing
Migrate the API code to use Gin for routing instead of using the plain
Go http module. This has several benefits: more correct header handling
(which also makes necessary an update of the Cypress tests as the
content type header is now correctly set to JSON), simpler code, safer
code (authentication done by a middleware which makes it harder to
forget), easier routing for a future v2 of the API, etc.