webui: Don't show API keys
API keys are considered secrets and should not simply be listed in plain
text in an easily accessible list. So do not send API keys to the user
and only show the key that has last been generated. This also
intentionally drops retrieving the keys from the database as they are
really never needed internally in any code either and therefore there is
no need to even have them in memory.

Also add a little dialog before creating a new key explaining the
"rules" of the key generation.

This commit also adds the React Bootstrap dependency, used for the modal
dialog.