webui: Simplify HTML templates
Move common code of all templates into the respective sub templates to
reduce the size of the templates. One purpose of this is to make making
changes to these parts easier - as done here with adding the lang
attribute to the html tag. Another purpose is to make the remaining
template code easier to analyse for futher simplification and future
removal.