common: Move the config file related code into a separate directory
This creates a new module "config" and moves the code related to parsing
the configuration file from the common module into there. The idea is to
start disentangling the code in the common module, establishing clearer
interfaces between the different parts and thus making it easier to
maintain. For the same reason this also moves the code for creating
database configuration objects from the config code to the database code
as it depends on the specific database driver. This way the config code
only has to deal with parsing and storing the toml config.