common: Move database connection code to a separate database module
Start moving the database related code into a separate module for easier
maintenance. This is still work in progress because the functions for
working with the logging and statistics tables as well as those for
working with the sqlite_databases table has not been migrated yet.

This might look like a major change but it is mostly moving around code
and splitting it up into several files. Actual changes here are limited
to these:

The code for creating the connection pool and for connecting to the job
queue is merged into one function.

Adding the default user and licences is now part of the connection code
and therefore does not need to be done in the frontend anymore.

Some sanitisation of log string has been removed. The database code
should not have to worry about this and the values here are already
validated by the frontend modules.

The default user gets an empty auth0 id instead of a random one. No
consequences to be expected from that because the default user is never
authenticated via auth0.

The code for toggling if a database is starred by a user is rewritten so
it does not use the databaseId helper function any longer.