api: Unify API call logging
This moves adding entries to the API call log to a Gin middleware in
order to simplify and unify the API logging code. The effect of this is
that there is one API specific API operation per API endpoint and that
logging always happens at the same time when processing an API call.
Because this also adds a 'v1' to the logged operations of v1 API calls,
a migration adds this for existing log records.