api: Add used API key to API call log
Add a new column to the API call log table for storing the id of the
API key used for the request. For adding a foreign key on this column
to the api_keys table, change the primary key of that table from the
key to the key_id column. The key column still has a unique index on it
for performance.