api, common: Add more information to API call log
When logging calls to the API endpoint include some additional useful
information: time it took to execute the handler, HTTP method of the
request, the returned HTTP status code, and the request and response
body sizes.

For some reason at least on my local system the response size is not
working correctly yet though.