Add maximum upload size to the usage limits table
Add a column to configure the maximum upload size to the usage_limits
database table. This allows different users to have different limits for
the maximum upload size.

Because it is now fully configurable the size override option in the
config file is removed as well. The users listed in there should get
assigned to a matching usage limit of course.