api: Ensure the caller receives a JSON response even for API access errors
This commit returns a JSON "error" string when an API authorisation
error occurs (ie missing API key, or an API key with insufficient access).

Prior to this, the only indication that something was wrong was through
the response header status code.

Was super confusing when using curl with jq to process responses, only
to have what looked like silent failures. ;)