api, webui: Clear cache usage limits when changing assigned limits
When changing the usage limits of a user via the web UI clear the cached
usage limits. This makes sure the newly assigned limits are loaded at
the time of the next API call. This means regular loading of usage data
from the database is no longer necessary for applying newly assigned
limits, so the time for reloading can be increased (or even reloading
completely removed). Note this only works when the api and the webui
backend share the same memcached instance.