minio: verify minio connection actually works
The MinIO "New()" function just verifies the connection information
*seems* like it should be ok, but doesn't actually connect to the
remote server and make sure things *are* working properly.

This commit adds a simple test of the MinIO connection to make sure
the connection is actually ok.  This is so connection problems
(firewall, routing, etc) are will show up when the service tries to
start (and fails) which our monitoring systems will catch and alert
us on.