api: Add preliminaries for API v2 endpoints
Add the code for authenticating and routing calls to future API v2
endpoints. Add a basic status endpoint for testing and a test for that.

Authentication works by either providing an API key or a session cookie.
The API key is intended for regular API use, the authentication by
session cookie is intended for calls from the web UI. This means that
the memcache instance used for session storage must be shared between
API and web UI.

For basic testing a "status" endpoint which returns ok when
authenticated successfully is added.