webui: Set cookie domain and set session cookie in dev environment
The API v2 endpoints allow requests from the web UI to be authenticated
by providing a session cookie. Because the web UI and the API can run on
different subdomains, set the domain attribute of the session cookie to
the top domain.

To properly develop and test API calls via the cookie authentication
also make sure the session cookie is also set when using the development
environment without logging in.