common: Fix connection string problem with gorm
Gorm doesn't seem to work with the 'pool_max_conns' parameter that
we use for the main PG connection pool, so this commit fixes that.

From quickly scanning the Gorm docs, it does seem to support
connection pools in some way.  But it might need a different
approach to get working (unsure personally).